Move slade3.cfg to a new slade.sqlite database file (experimental)
Just a bit of an experiment to see if something like this is viable, at least for user settings / configuration. I'd like to eventually implement some kind of archive 'library' which would use a database since text files can get a bit unwieldy with lots of data.

So far works pretty well it looks like, with an added bonus of not losing settings etc if the program exits abnormally (ie. crashes :P)